pyproject.toml
[project]
name = "lipcoding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.3",
    "django-extensions>=4.1",
    "django-ninja>=1.4.3",
    "pytest-django>=4.11.1",
    "ruff>=0.12.1",
]
